ft_printf V2

+++ ft_printf/ft_putstr.c
@@ -6,7 +6,7 @@
 /*   By: eguelin <eguelin@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/19 11:22:47 by eguelin           #+#    #+#             */
-/*   Updated: 2022/11/20 11:50:42 by eguelin          ###   ########lyon.fr   */
+/*   Updated: 2022/11/22 20:05:48 by eguelin          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_putstr(const char *s)
 {
-	size_t	pos;
+	int	pos;
 
 	pos = 0;
 	while (s[pos])
